#28 - added redirect uri in token endpoint - refactor connect controller - fixed tests
