+++ server/src/Sven/Models/AuthCode.cs
@@ -7,7 +7,6 @@ namespace Sven.Models
         public string Code { get; set; } = string.Empty;
         public string CodeChallenge { get; set; } = string.Empty;
         public string CodeChallengeMethod { get; set; } = AuthConstants.OAuth.CodeChallengeMethods.Sha256;
-        public string RedirectUri { get; set; } = string.Empty;
         public DateTimeOffset ExpiresAt { get; set; }
     }
 }
