+++ server/tests/Sven.Tests/Services/TokenProviderTests.cs
@@ -18,6 +18,7 @@ namespace Sven.Tests.Services
         private readonly SvenTokenProvider _provider;
 
         private readonly string _clientId = "client1";
+        private readonly string _redirectUri = "http://localhost";
 
         public TokenProviderTests()
         {
@@ -40,6 +41,7 @@ namespace Sven.Tests.Services
             RefreshToken validToken = new RefreshToken()
             {
                 ClientId = _clientId,
+                RedirectUri = _redirectUri,
                 Scope = $"{AuthConstants.Scopes.OpenId} {AuthConstants.Scopes.Profile}",
                 ExpiresAt = DateTime.UtcNow.AddHours(1)
             };
@@ -47,7 +49,7 @@ namespace Sven.Tests.Services
                 .Returns(new Result<RefreshToken>(validToken));
 
             Result<bool> result = await _provider.IsRefreshTokenValidAsync(
-                "valid_token", _clientId, AuthConstants.Scopes.OpenId, CancellationToken.None);
+                "valid_token", _clientId, _redirectUri, AuthConstants.Scopes.OpenId, CancellationToken.None);
 
             Assert.True(result.IsSuccess);
         }
@@ -61,7 +63,7 @@ namespace Sven.Tests.Services
                 .Returns(new Result<RefreshToken>(resultErrorExpected));
 
             Result<bool> result = await _provider.IsRefreshTokenValidAsync(
-                "missing_token", _clientId, AuthConstants.Scopes.OpenId, CancellationToken.None);
+                "missing_token", _clientId, _redirectUri, AuthConstants.Scopes.OpenId, CancellationToken.None);
 
             Assert.True(result.IsError);
             Assert.Equal(AuthConstants.OAuth.Errors.InvalidGrant, result.Error.ErrorMessage);
@@ -76,6 +78,7 @@ namespace Sven.Tests.Services
             RefreshToken token = new RefreshToken()
             {
                 ClientId = _clientId,
+                RedirectUri = _redirectUri,
                 Scope = AuthConstants.Scopes.OpenId,
                 ExpiresAt = DateTime.UtcNow.AddHours(1)
             };
@@ -84,7 +87,7 @@ namespace Sven.Tests.Services
                 .Returns(new Result<RefreshToken>(token));
 
             Result<bool> result = await _provider.IsRefreshTokenValidAsync(
-                "token", _clientId, $"{AuthConstants.Scopes.OpenId} {AuthConstants.Scopes.Profile}", CancellationToken.None);
+                "token", _clientId, _redirectUri, $"{AuthConstants.Scopes.OpenId} {AuthConstants.Scopes.Profile}", CancellationToken.None);
 
             Assert.True(result.IsError);
             Assert.Equal(AuthConstants.OAuth.Errors.InvalidScope, result.Error.ErrorMessage);
@@ -97,6 +100,7 @@ namespace Sven.Tests.Services
             RefreshToken token = new RefreshToken()
             {
                 ClientId = _clientId,
+                RedirectUri = _redirectUri,
                 Scope = AuthConstants.Scopes.OpenId,
                 ExpiresAt = DateTime.UtcNow.AddHours(1)
             };
@@ -104,7 +108,28 @@ namespace Sven.Tests.Services
                 .Returns(new Result<RefreshToken>(token));
 
             Result<bool> result = await _provider.IsRefreshTokenValidAsync(
-                "token", "wrong_client", AuthConstants.Scopes.OpenId, CancellationToken.None);
+                "token", "wrong_client", _redirectUri, AuthConstants.Scopes.OpenId, CancellationToken.None);
+
+            Assert.True(result.IsError);
+            Assert.Equal(AuthConstants.OAuth.Errors.InvalidClient, result.Error.ErrorMessage);
+        }
+
+        [Fact]
+        [Trait("Category", "Unit")]
+        public async Task IsRefreshTokenValidAsync_RedirectUriMismatch_ReturnsInvalidClient()
+        {
+            RefreshToken token = new RefreshToken()
+            {
+                ClientId = _clientId,
+                RedirectUri = _redirectUri,
+                Scope = AuthConstants.Scopes.OpenId,
+                ExpiresAt = DateTime.UtcNow.AddHours(1)
+            };
+            _refreshTokenStore.GetAsync("token", Arg.Any<CancellationToken>())
+                .Returns(new Result<RefreshToken>(token));
+
+            Result<bool> result = await _provider.IsRefreshTokenValidAsync(
+                "token", _clientId, "https://wronguri", AuthConstants.Scopes.OpenId, CancellationToken.None);
 
             Assert.True(result.IsError);
             Assert.Equal(AuthConstants.OAuth.Errors.InvalidClient, result.Error.ErrorMessage);
@@ -117,6 +142,7 @@ namespace Sven.Tests.Services
             RefreshToken token = new RefreshToken()
             {
                 ClientId = _clientId,
+                RedirectUri = _redirectUri,
                 Scope = AuthConstants.Scopes.OpenId,
                 ExpiresAt = DateTime.UtcNow.AddHours(-1)
             };
@@ -125,7 +151,7 @@ namespace Sven.Tests.Services
             _timeProvider.GetUtcNow().Returns(DateTime.UtcNow);
 
             Result<bool> result = await _provider.IsRefreshTokenValidAsync(
-                "token", _clientId, AuthConstants.Scopes.OpenId, CancellationToken.None);
+                "token", _clientId, _redirectUri, AuthConstants.Scopes.OpenId, CancellationToken.None);
 
             Assert.True(result.IsError);
             Assert.Equal(AuthConstants.OAuth.Errors.InvalidGrant, result.Error.ErrorMessage);
